Add files via upload
Simple and User-Friendly Interface 🌟
I designed the password generator with a straightforward command-line interface, making it accessible to all levels of users. Running the script is as easy as typing a few commands in the terminal!

Customizable Password Criteria 🔤
I wanted my password generator to be versatile, so I made sure to include various criteria to customize the generated passwords. You can choose the length of the passwords and decide whether to include uppercase letters, lowercase letters, numbers, and special symbols.

Robust Randomness 🎲
The key to a secure password is randomness. To ensure this, I leveraged Python's built-in random module to shuffle the characters and create passwords that are incredibly difficult to guess.

No More Reusing Passwords! ♻️
One of the essential features I integrated into the generator is the ability to create multiple unique passwords in one go. This discourages the common (and dangerous!) habit of reusing passwords across different accounts.

Security First! 🔒
As an added security measure, I made sure that the generated passwords do not contain easily confused characters, such as '1' and 'l' or 'O' and '0'. This minimizes the risk of human error when typing passwords.

Open-Source Goodness 📜
I'm a big fan of open-source software, so I decided to share the source code of my password generator on GitHub. Feel free to check it out, suggest improvements, or even contribute to the project!